+++ nctest/slabs.c
@@ -157,7 +157,7 @@ test_slabs(cdfid)
     /* define a multi-dimensional variable of each type */
 
     for (iv = 0; iv < NVARS; iv++) {
-	va[iv].dims = (int *) emalloc(sizeof(int) * va[iv].ndims);
+	va[iv].dims = (int *) emalloc(sizeof(int) * (size_t)va[iv].ndims);
 	for (idim = 0; idim < va[iv].ndims; idim++)
 	  va[iv].dims[idim] = dimids[idim];
 	varid[iv] = ncvardef(cdfid, va[iv].name, va[iv].type, va[iv].ndims,
@@ -177,7 +177,7 @@ test_slabs(cdfid)
 
     for (iv = 0; iv < NVARS; iv++) { /* test each type of variable */
 
-	v = emalloc(WSIZE*XSIZE*YSIZE*ZSIZE * nctypelen(va[iv].type));
+	v = emalloc(WSIZE*XSIZE*YSIZE*ZSIZE * (size_t)nctypelen(va[iv].type));
 
 	/* fill it with values using a function of dimension indices */
 	ii = 0;
