+++ nctest/vartests.c
@@ -48,7 +48,7 @@ test_ncvarid(path)
 	ncclose(cdfid); return ++nerrs;
     }
     /* in define mode, add a variable */
-    xx.dims = (int *) emalloc(sizeof(int) * xx.ndims);
+    xx.dims = (int *) emalloc(sizeof(int) * (size_t)xx.ndims);
     for (id = 0; id < xx.ndims; id++)
       xx.dims[id] = id;
     if ((varid = ncvardef(cdfid,
@@ -584,7 +584,7 @@ test_ncvarrename(path)
 	ncclose(cdfid); return ++nerrs;
     }
     /* in define mode, add two variables */
-    yy.dims = (int *) emalloc(sizeof(int) * yy.ndims);
+    yy.dims = (int *) emalloc(sizeof(int) * (size_t)yy.ndims);
     for (id = 0; id < yy.ndims; id++)
       yy.dims[id] = id;
     if ((yy_id = ncvardef(cdfid,
@@ -593,7 +593,7 @@ test_ncvarrename(path)
 	ncclose(cdfid); return ++nerrs;
     }
     add_var(&test, &yy);	/* keep in-memory netcdf in sync */
-    zz.dims = (int *) emalloc(sizeof(int) * zz.ndims);
+    zz.dims = (int *) emalloc(sizeof(int) * (size_t)zz.ndims);
     for (id = 0; id < zz.ndims; id++)
       zz.dims[id] = id;
     if (ncvardef(cdfid, zz.name, zz.type, zz.ndims, zz.dims) == -1) {
